Refactor probe generation and handling for fork-related functions
- Updated the `probes-haridev-lead.json` to include fork-related functions in the libc probe list.
- Modified `test_dlio_benchmark.sh` to streamline the usage of LD_PRELOAD.
- Removed hardcoded probe names for fork and vfork from `generator.cpp` to simplify event updates.
- Enhanced `kprobe_generator.h`, `syscall_generator.h`, and `uprobe_generator.h` to differentiate exit functions for fork-related probes.
- Introduced `generic_fork_exit` function in `common.h` to handle tracing for fork events.
- Cleaned up `init.bpf.c` by removing redundant fork and vfork syscall probes, relying on the new generic exit handling.
