+++ etc/datacrumbs/data/probes-haridev-lead.json
@@ -166,197 +166,25 @@
     "type":2,
     "name":"libc",
     "functions":[
-      "_Unwind_Resume",
-      "_authenticate@GLIBC_2.2.5",
-      "_fitoa_word",
-      "_help",
-      "_init_first",
-      "_int_memalign",
-      "_longjmp_unwind",
-      "_nl_find_domain",
-      "_nl_finddomain_subfreeres",
-      "_nl_load_domain",
-      "_nl_make_l10nflist",
-      "_nl_normalize_codeset",
-      "_no_syscall",
-      "_quicksort",
-      "action_list_add__",
-      "add_argless_short_opt",
-      "add_pad",
-      "advance@GLIBC_2.2.5",
-      "aio_cancel64@@GLIBC_2.34",
-      "aio_cancel64@GLIBC_2.2.5",
-      "aio_cancel@@GLIBC_2.34",
-      "aio_cancel@GLIBC_2.2.5",
-      "aio_error64@@GLIBC_2.34",
-      "aio_error64@GLIBC_2.2.5",
-      "aio_error@@GLIBC_2.34",
-      "aio_error@GLIBC_2.2.5",
-      "aio_fsync64@GLIBC_2.2.5",
-      "aio_fsync@@GLIBC_2.34",
-      "aio_fsync@GLIBC_2.2.5",
-      "aio_init@@GLIBC_2.34",
-      "aio_init@GLIBC_2.2.5",
-      "aio_read64@@GLIBC_2.34",
-      "aio_read64@GLIBC_2.2.5",
-      "aio_read@@GLIBC_2.34",
-      "aio_read@GLIBC_2.2.5",
-      "aio_return64@@GLIBC_2.34",
-      "aio_return64@GLIBC_2.2.5",
-      "aio_return@@GLIBC_2.34",
-      "aio_return@GLIBC_2.2.5",
-      "aio_suspend64@@GLIBC_2.34",
-      "aio_suspend64@GLIBC_2.2.5",
-      "aio_suspend@@GLIBC_2.34",
-      "aio_suspend@GLIBC_2.2.5",
-      "aio_write64@@GLIBC_2.34",
-      "aio_write64@GLIBC_2.2.5",
-      "aio_write@@GLIBC_2.34",
-      "aio_write@GLIBC_2.2.5",
-      "arg",
-      "arg_trimdomain_list",
-      "argp_args_usage",
-      "argp_version_parser",
-      "authdes_validate",
-      "authnone_destroy",
-      "authnone_marshal",
-      "authnone_refresh",
-      "backtrace_helper",
-      "build_wcs_buffer",
-      "calc_sizes",
-      "call_once@@GLIBC_2.34",
-      "call_once@GLIBC_2.28",
-      "callback",
-      "cbc_crypt@GLIBC_2.2.5",
-      "check_stdfiles_vtables",
-      "clone3",
-      "comma",
-      "conf_decrement.part.0",
-      "copysignf64",
-      "create_module@GLIBC_2.2.5",
-      "day_of_the_week",
-      "deallocate",
-      "do_aio_misc_wait",
-      "do_futex_wait",
-      "do_init",
-      "do_sym",
-      "error_tail",
-      "eval_expr",
-      "eval_expr_val",
-      "fclose@@GLIBC_2.2.5",
-      "fdopen@@GLIBC_2.2.5",
-      "fgetpos64@@GLIBC_2.2.5",
-      "fgetpos@@GLIBC_2.2.5",
-      "flush_cleanup",
-      "flush_out",
-      "fnmatch_convert_to_wide",
-      "fopen@@GLIBC_2.2.5",
-      "free_derivation",
-      "free_modules_db",
-      "free_slotinfo",
-      "frexpf32x",
-      "frexpf64",
-      "frexpf64x",
-      "fsetpos64@@GLIBC_2.2.5",
-      "fsetpos@@GLIBC_2.2.5",
-      "fts_safe_changedir.part.0",
-      "gaiconf_init",
-      "gconv_parse_code",
-      "get_input_bytes",
-      "getaliasent_r@@GLIBC_2.2.5",
-      "gethostbyname3_multi",
-      "glob64@@GLIBC_2.27",
-      "glob@@GLIBC_2.27",
-      "global_state_allocate",
-      "hol_find_entry.isra.0",
-      "hol_usage",
-      "init",
-      "int_mallinfo",
-      "internal_fallocate",
-      "iruserfopen",
-      "isastream@GLIBC_2.2.5",
-      "known_compare",
-      "ldexpf128",
-      "ldexpf32x",
-      "ldexpf64x",
-      "lio_listio64@GLIBC_2.2.5",
-      "lio_listio64@GLIBC_2.4",
-      "lio_listio@@GLIBC_2.34",
-      "lio_listio@GLIBC_2.2.5",
-      "lio_listio@GLIBC_2.4",
-      "lio_listio_internal",
-      "llseek@GLIBC_2.2.5",
-      "logout@GLIBC_2.2.5",
-      "logwtmp@GLIBC_2.2.5",
-      "malloc_printerr",
-      "marshal_new_auth.isra.0",
-      "memcpy@GLIBC_2.2.5",
-      "mmap_remap_check",
-      "modff32",
-      "modff32x",
-      "modff64",
-      "modff64x",
-      "module_load",
-      "module_load_builtin",
-      "new_heap",
-      "notification_function",
-      "opendir_tail",
-      "openlog_internal",
-      "openpty@GLIBC_2.2.5",
-      "parse_arith",
-      "parse_qtd_backslash",
-      "parse_tilde",
-      "parse_tzname",
-      "pclose@@GLIBC_2.2.5",
-      "plural_eval",
-      "pmap_set@GLIBC_2.2.5",
-      "popen@@GLIBC_2.2.5",
-      "posix_spawn@@GLIBC_2.15",
-      "posix_spawn@GLIBC_2.2.5",
-      "posix_spawnp@@GLIBC_2.15",
-      "posix_spawnp@GLIBC_2.2.5",
-      "print_errno_message",
-      "profil_count",
-      "putpmsg@GLIBC_2.2.5",
-      "re_node_set_init_copy",
-      "read_last_entry",
-      "realpath@@GLIBC_2.3",
-      "realpath@GLIBC_2.2.5",
-      "regexec@@GLIBC_2.3.4",
-      "regexec@GLIBC_2.2.5",
-      "res_hnok@@GLIBC_2.34",
-      "res_ownok@@GLIBC_2.34",
-      "rfc3484_sort",
-      "ruserok2_sa",
-      "save_for_backup",
-      "save_for_wbackup",
-      "scalbnf32x",
-      "sched_getaffinity@GLIBC_2.3.3",
-      "scopecmp",
-      "search_list_add__",
-      "shm_open@@GLIBC_2.34",
-      "shm_unlink@@GLIBC_2.34",
-      "shm_unlink@GLIBC_2.2.5",
-      "shmctl@@GLIBC_2.2.5",
-      "step@GLIBC_2.2.5",
-      "stime@GLIBC_2.2.5",
-      "strchr_ifunc",
-      "strip_whitespace",
-      "strlist_add__",
-      "tcache_init.part.0",
-      "tmpfile@@GLIBC_2.2.5",
-      "trecurse",
-      "trecurse_r",
-      "try",
-      "try_file_lock",
-      "universal",
-      "unwind_cleanup",
-      "ustat@GLIBC_2.2.5",
-      "vtimes@GLIBC_2.2.5",
-      "w_addchar",
-      "write_gmon",
-      "write_string",
-      "x_inline"
+      "__GI___fork",
+      "__GI___register_atfork",
+      "__GI___vfork",
+      "__GI_forkpty",
+      "__dyn_pthread_atfork",
+      "__forkpty",
+      "__libc_unwind_link_after_fork",
+      "__libc_vfork",
+      "__malloc_fork_lock_parent",
+      "__malloc_fork_unlock_child",
+      "__malloc_fork_unlock_parent",
+      "__mq_notify_fork_subprocess",
+      "__nss_database_fork_prepare_parent",
+      "__nss_database_fork_subprocess",
+      "__run_postfork_handlers",
+      "__run_prefork_handlers",
+      "__timer_fork_subprocess",
+      "__unregister_atfork",
+      "pthread_atfork@GLIBC_2.2.5"
     ],
     "binary_path":"\/usr\/lib64\/libc.so.6",
     "function_offsets":[
