+++ src/datacrumbs/generator/mechanisms/syscall_generator.h
@@ -3,9 +3,9 @@
 #include <datacrumbs/common/logging.h>  // Logging macros
 
 #include <algorithm>
+#include <regex>
 #include <sstream>
 #include <string>
-
 namespace datacrumbs {
 
 /**
@@ -46,10 +46,17 @@ class SyscallGenerator {
     ss << "  return 0;\n";
     ss << "}\n";
     // Generate return probe
+    bool is_fork = std::regex_search(func_name_, std::regex(".*fork.*"));
+    std::string exit_func;
+    if (is_fork) {
+      exit_func = "generic_fork_exit";
+    } else {
+      exit_func = "generic_exit";
+    }
     ss << "SEC(\"kretsyscall/" << func_name_ << "\")\n";
     ss << "int BPF_KRETPROBE(" << sanitized_func_name << event_id_
        << "_exit, struct pt_regs* regs) {\n";
-    ss << "  generic_exit(ctx, " << event_id_ << ");\n";
+    ss << "  " << exit_func << "(ctx, " << event_id_ << ");\n";
     ss << "  return 0;\n";
     ss << "}\n";
