+++ src/datacrumbs/server/bpf/common.h
@@ -35,8 +35,7 @@ static inline __attribute__((always_inline)) u32 hash_str(const char* str, size_
   return hash;
 }
 
-static inline __attribute__((always_inline)) u32 hash_and_store(struct string_t *str,
-                                                                size_t len) {
+static inline __attribute__((always_inline)) u32 hash_and_store(struct string_t* str, size_t len) {
   u32* existing = bpf_map_lookup_elem(&file_map, str);
   if (existing) {
     return *existing;  // Return existing hash
@@ -154,12 +153,12 @@ static inline __attribute__((always_inline)) int usdt_exit(struct pt_regs* ctx, 
   event->id = key.id;
   event->event_id = event_id;
   DATACRUMBS_COLLECT_TIME(event);
-  struct string_t local_str = {};                                            // 100
+  struct string_t local_str = {};                                                      // 100
   long len = bpf_probe_read_user_str(&local_str.str, MAX_STR_READ_LEN, (void*)clazz);  // 90
   u32 class_hash = hash_and_store(&local_str, len);
-  event->class_hash = class_hash;                                              // 100
+  event->class_hash = class_hash;                                                  // 100
   len = bpf_probe_read_user_str(&local_str.str, MAX_STR_READ_LEN, (void*)method);  // 90
-  
+
   u32 method_hash = hash_and_store(&local_str, len);
   event->method_hash = method_hash;  // 100
   DATACRUMBS_EVENT_SUBMIT(event);
@@ -172,5 +171,32 @@ static inline __attribute__((always_inline)) int usdt_exit(struct pt_regs* ctx, 
   return 0;
 }
 #endif
+
+#if defined(DATACRUMBS_TRACING_ENABLE) && (DATACRUMBS_TRACING_ENABLE == 1)
+static inline __attribute__((always_inline)) int generic_fork_exit(struct pt_regs* ctx,
+                                                                   u64 event_id) {
+  struct fn_key_t key = {};
+  key.event_id = event_id;
+  u64 start_ts;
+  if (need_tracing(&key, &start_ts)) {
+    // u64 id = bpf_get_current_pid_tgid();
+    u64 tsp = bpf_ktime_get_ns();
+    u32 pid = PT_REGS_RC(ctx);
+    (void)pid;
+    if (pid != 0) {
+      DBG_PRINTK("Collect forked tracing PID %d", pid);
+      bpf_map_update_elem(&pid_map, &pid, &tsp, BPF_ANY);
+    }
+  }
+  return generic_exit(ctx, event_id);
+}
+
+#else
+static inline __attribute__((always_inline)) int generic_fork_exit(struct pt_regs* ctx,
+                                                                   u64 event_id) {
+  return 0;
+}
+#endif
+
 char LICENSE[] SEC("license") = "Dual BSD/GPL";
 #endif  // DATACRUMBS_SERVER_BPF_COMMON_H
